requirements.txt
opencv-python
numpy
pillow
pyautogui
pywin32
paddlepaddle
paddleocr
pydub
pygame